Modified the SQL Table and also the python script for accumulating stats for batters (have to do same for pitchers soon)
Instead of pulling the aggregated stats for rbi, ibb, war, and runs scored I instead pull it straight from the fangraphs API so I can get it on a game level. I have to do similar for pitching but for now that data is fully aggregated. This allows for users to filter these stats like WAR, RBI, and IBB over a date span as well as for a full season